requirements.txt
torch==2.0.1
torchvision==0.15.2
Pillow==10.0.0
matplotlib==3.7.1
